% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTprocessR_S3.R
\name{autoplot.bt_stones}
\alias{autoplot.bt_stones}
\title{An autoplot method for the class "bt_stones". Plots a bayes factor matrix.}
\usage{
\method{autoplot}{bt_stones}(x)
}
\description{
An autoplot method for the class "bt_stones". Plots a bayes factor matrix.
}
