% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{branchColours}
\alias{branchColours}
\title{branchColours
Generates the edge colours to colour edges by total rate.}
\usage{
branchColours(PP, opts)
}
\description{
branchColours
Generates the edge colours to colour edges by total rate.
}
\keyword{internal}
