% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPosts.R
\name{compPosts}
\alias{compPosts}
\title{compPosts}
\usage{
compPosts(logs, parameter, alpha = 0.7, ...)
}
\arguments{
\item{logs}{A vector of EITHER two or more logfile names, OR a list of two 
or more objects of class bt_post (see \link[BTprocessR]{loadPosterior}).}

\item{parameter}{A vector containing the names of the parameter to compare.}

\item{...}{Additional arguments passed to \link[BTprocessR]{loadPosterior},
in the case the filenames are given for logs.}
}
\description{
Compares histograms of one or more parameters two or more output files.
Generates a plot showing the distributions of the same parameter from two
or more posterior samples from BayesTraits MCMC.
}
