% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{generateTrans}
\alias{generateTrans}
\title{generateTrans
Generate the transparencies from data and colours}
\usage{
generateTrans(cols, values)
}
\arguments{
\item{cols}{The colours to make transparent}

\item{values}{The data to set transparency by}
}
\description{
generateTrans
Generate the transparencies from data and colours
}
\keyword{internal}
