% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{getPalette}
\alias{getPalette}
\alias{generateTrans}
\title{getPalette
Generate the palette for colour ramps.}
\usage{
getPalette(opts)

generateTrans(cols, values)
}
\description{
getPalette
Generate the palette for colour ramps.

generateTrans
Generate the transparencies from data and colours
}
\keyword{internal}
