% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{getPostStats}
\alias{getPostStats}
\title{getPostStats}
\usage{
getPostStats(logfile, parameter, ...)
}
\arguments{
\item{logfile}{The name of the logfile of the BayesTraits analysis, or an
object of class "bt_post" (see \link[BTprocessR]{loadPosterior})}

\item{parameter}{The name(s) of one or more parameters present in the logfile.}

\item{...}{Additional arguments passed to \link[BTprocessR]{loadPosterior}.}
}
\value{
An object of class data.frame with a row for each parameter, and
columns for the mean, median, mode and standard deviation of those
parameters.
}
\description{
Given a posterior of MCMC samples from a BayesTraits analysis (a logfile -
typically with the .log.txt extension) and the name of one or morea
parameters present in that logfile this will return the mean, median, mode
and standard deviation of the parameter estimates.
}
\examples{
params <- getParams("cool-data.log.txt")
getPostMean("cool-data.log.txt", params[c(2:5)]
getPostMean("cool-data.log.txt", "Lh")
}
