% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalLikelhoods.R
\name{getStones}
\alias{getStones}
\title{getStones}
\usage{
getStones(stones, labels = NULL)
}
\arguments{
\item{stones}{A vector of one or more log files from stepping stones 
analysis.}

\item{labels}{A vector of equal length to the number of stones files 
containing a name for each model. If left NULL then the filenames will be 
used (which may be long and unwieldy when plotting).}
}
\value{
A data.frame with a row per stones file, and a column with the
marginal likelihoods of each model in.
}
\description{
Get the marginal likelihoods from from one or more stepping stones log
files.
}
