% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{legendInfo}
\alias{legendInfo}
\title{legendInfo
Generates legend info for colour scale legends}
\usage{
legendInfo(tree, opts, cols)
}
\description{
legendInfo
Generates legend info for colour scale legends
}
\keyword{internal}
