% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btSetupFuncs.R
\name{makeInfile}
\alias{makeInfile}
\title{makeInfile
Takes a tree and some data, ladderizes the tree, matches the tree and the
data and then makes a run file based on a control list. The prepared tree,
data, infile and command to execute the analysis will be written out to
the analysis_path with the prefixes analysis_name.
Sets up basic
BayesTraits analysis - for more complex analyses (e.g. alternate priors,
parameter restriction, fixed scalars on nodes, fossilised states etc.) users 
must write their own control files.}
\usage{
makeInfile(tree, data, species_col = NULL, analysis_name, analysis_path,
  job = FALSE, bt.control = list(), job.control = list())
}
\arguments{
\item{tree}{The tree for the analysis to be run on.}

\item{data}{The data to analyse}

\item{analysis_path}{The file path to put the resultant files in - defaults
to working directory.}

\item{analaysis_name}{The name for the analysis - will be prefixed to all
files generated.}

\item{control}{List of options for BayesTraits.}
}
\description{
makeInfile
Takes a tree and some data, ladderizes the tree, matches the tree and the
data and then makes a run file based on a control list. The prepared tree,
data, infile and command to execute the analysis will be written out to
the analysis_path with the prefixes analysis_name.
Sets up basic
BayesTraits analysis - for more complex analyses (e.g. alternate priors,
parameter restriction, fixed scalars on nodes, fossilised states etc.) users 
must write their own control files.
}
