% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeStructureFunctions.R
\name{makeKey}
\alias{makeKey}
\title{sharedBranches}
\usage{
makeKey(tree)
}
\arguments{
\item{tree}{An object of class "phylo"}
}
\description{
Make species key from a tree.
}
\keyword{internal}
