% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{makeLegendTitle}
\alias{makeLegendTitle}
\title{makeLegendTitle
Generates legend title for scale bars.}
\usage{
makeLegendTitle(opts)
}
\description{
makeLegendTitle
Generates legend title for scale bars.
}
\keyword{internal}
