% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjpp.R
\name{makeSpKey}
\alias{makeSpKey}
\title{makeSpKey
Detaches the species descendant from a node from the counts table and adds an
id column to match to a species key list.}
\usage{
makeSpKey(counts)
}
\arguments{
\item{counts}{The finished counts table.}
}
\description{
makeSpKey
Detaches the species descendant from a node from the counts table and adds an
id column to match to a species key list.
}
\keyword{internal}
