% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{nodeShapes}
\alias{nodeShapes}
\title{nodeShapes
Generates the node labels, edge.colours and transparencies to plot the
location of transformations in a posterior.}
\usage{
nodeShapes(PP, opts, mode)
}
\description{
nodeShapes
Generates the node labels, edge.colours and transparencies to plot the
location of transformations in a posterior.
}
\keyword{internal}
