% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTprocessR_S3.R
\name{plot.ancstates}
\alias{plot.ancstates}
\title{a print method for class ancstates.}
\usage{
\method{plot}{ancstates}(x, plot.options = list(), ...)
}
\description{
a print method for class ancstates.
}
