% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTprocessR_S3.R
\name{plot.bt_stones}
\alias{plot.bt_stones}
\title{A plot method for the class "bt_stones" that invokes the
autoplot method.}
\usage{
\method{plot}{bt_stones}(x)
}
\description{
A plot method for the class "bt_stones" that invokes the
autoplot method.
}
