% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTprocessR_S3.R
\name{plot.trees_summary}
\alias{plot.trees_summary}
\title{A plot method for the class "trees_summary" - the default will
plot all trees in a 2x2 plot, otherwise a tree(s) can be specified
to plot alongside the time-tree.}
\usage{
\method{plot}{trees_summary}(x, tree = NULL, tips = FALSE)
}
\description{
A plot method for the class "trees_summary" - the default will
plot all trees in a 2x2 plot, otherwise a tree(s) can be specified
to plot alongside the time-tree.
}
