% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{plotShapes}
\alias{plotShapes}
\title{plotShapes
Generates the node labels, edge.colours and transparencies to plot the
location of transformations in a posterior.}
\usage{
plotShapes(PP, opts, mode)
}
\description{
plotShapes
Generates the node labels, edge.colours and transparencies to plot the
location of transformations in a posterior.
}
\keyword{internal}
