% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTprocessR_S3.R
\name{print.bt_post}
\alias{print.bt_post}
\title{A print method for class bt_post
Displays some descriptive info about the posterior in a table}
\usage{
\method{print}{bt_post}(x)
}
\description{
A print method for class bt_post
Displays some descriptive info about the posterior in a table
}
