% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTprocessR_S3.R
\name{print.trees_summary}
\alias{print.trees_summary}
\title{Print function for the S3 class trees_summary}
\usage{
\method{print}{trees_summary}(x)
}
\description{
Print function for the S3 class trees_summary
}
