% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjpp.R
\name{scalarSearch}
\alias{scalarSearch}
\title{scalarSearch
Searches through the posterior of an RJ continuous model for scalars and 
returns them.}
\usage{
scalarSearch(rj_output, counts, fullmrcas, verbose)
}
\arguments{
\item{rj_output}{partially processed RJ output.}

\item{counts}{The counts table.}

\item{fullmrcas}{Most recent common ancestor tables for each tip in the tree,
generated internally.}
}
\description{
scalarSearch
Searches through the posterior of an RJ continuous model for scalars and 
returns them.
}
\keyword{internal}
