% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{shapeCols}
\alias{shapeCols}
\title{shapeCols
Generates colours for shapes for either node or branch scalars.}
\usage{
shapeCols(opts, plot.dat, mode)
}
\description{
shapeCols
Generates colours for shapes for either node or branch scalars.
}
\keyword{internal}
