% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeStructureFunctions.R
\name{sharedBranches}
\alias{sharedBranches}
\title{sharedBranches}
\usage{
sharedBranches(tree1, tree2)
}
\arguments{
\item{tree1}{An object of class "phylo"}

\item{tree2}{An object of class "phylo"}
}
\description{
Finds shared branches between two trees where the smaller tree is a subset of
the larger tree.
}
