% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjpp.R
\name{summariseRjpp}
\alias{summariseRjpp}
\title{summariseRjpp
This functions takes the somewhat massive output of the rjpp function and
pares it down to the scalars and/or rates the user is interested in.
QUESTION: Should this be limited to a single scalar type? That would include
the rates and the origins?
QUESTION: Is threshold important here? Or could that be in the autoplot
method?}
\usage{
summariseRjpp(PP, scalar)
}
\arguments{
\item{scalar}{The scalar to summarise the results of. Either:
node_scalar, branch_scalar, rate, lambda, delta, kappa or node_branch}

\item{pp}{An object of class "rjpp" - typically the output of the rjpp 
function.}

\item{threshold}{The probability threshold over which scalars will be show.
When equal to zero ALL scalars in the posterior will be returned. When equal 
to 0.5 only scalars present in greater than 50% of posterior samples will be 
returned, and so on.}
}
\description{
summariseRjpp
This functions takes the somewhat massive output of the rjpp function and
pares it down to the scalars and/or rates the user is interested in.
QUESTION: Should this be limited to a single scalar type? That would include
the rates and the origins?
QUESTION: Is threshold important here? Or could that be in the autoplot
method?
}
