% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjpp.R
\name{createCounts}
\alias{createCounts}
\alias{createCountsTable}
\title{createCounts
Creates the counts table for the rjpp.}
\usage{
createCountsTable(reftree, tree_summary)
}
\arguments{
\item{extree}{The time tree}

\item{meanbl}{The output of meanBranches}
}
\description{
createCounts
Creates the counts table for the rjpp.
}
\keyword{(e.g.}
\keyword{If}
\keyword{Is}
\keyword{NOTES}
\keyword{Take}
\keyword{This}
\keyword{a}
\keyword{a)}
\keyword{and}
\keyword{are}
\keyword{at}
\keyword{b)}
\keyword{be}
\keyword{being}
\keyword{best}
\keyword{branchlength}
\keyword{built}
\keyword{c)}
\keyword{can}
\keyword{d)}
\keyword{descriptive}
\keyword{for}
\keyword{headings}
\keyword{here,}
\keyword{if:}
\keyword{in}
\keyword{information,}
\keyword{information?)}
\keyword{internal}
\keyword{into}
\keyword{is}
\keyword{it}
\keyword{later}
\keyword{look}
\keyword{makes}
\keyword{massive}
\keyword{more}
\keyword{occur}
\keyword{on.}
\keyword{one}
\keyword{or}
\keyword{populated}
\keyword{possible}
\keyword{rates}
\keyword{see}
\keyword{sensible}
\keyword{split}
\keyword{splits,}
\keyword{table}
\keyword{than}
\keyword{that}
\keyword{the}
\keyword{there}
\keyword{this}
\keyword{unwieldy}
\keyword{useful}
\keyword{way?}
