% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjpp.R
\name{createCounts}
\alias{createCounts}
\alias{createCountsTable}
\title{createCounts
Creates the counts table for the rjpp.}
\usage{
createCountsTable(reftree, tree_summary)
}
\arguments{
\item{reftree}{The time tree}

\item{tree_summary}{The output of summariseTrees}
}
\description{
createCounts
Creates the counts table for the rjpp.
}
\keyword{internal}
