% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalLikelhoods.R, R/summaryFunctions.R
\name{getStones}
\alias{getStones}
\alias{getStones}
\title{getStones}
\usage{
getStones(stonesfile, order = TRUE)

getStones(stonesfile, order = TRUE)
}
\arguments{
\item{stonesfile}{A vector of one or more log files from stepping stones analysis.}

\item{order}{If TRUE then the resultant data.frame is ordered according to
marginal likelihood.}

\item{stones}{A vector of one or more log files from stepping stones analysis.}

\item{labels}{A vector of equal length to the number of stones files containing a name for each model. If left NULL then the filenames will be used (which may be long and unwieldy when plotting).}
}
\value{
A data.frame with a row per stones file, and a column with the
marginal likelihoods of each model in.

A data.frame with a row per stones file, and a column with the
marginal likelihoods of each model in.
}
\description{
Get the marginal likelihoods from from one or more stepping stones log
files.

Get the marginal likelihoods from from one or more stepping stones log
files.
}
