% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{plotShifts}
\alias{plotShifts}
\title{plotShifts}
\usage{
plotShifts(PP, plot.options = list(), ...)
}
\arguments{
\item{PP}{The output of the rjpp function.}

\item{plot.options}{A list of control options. See description.}
}
\description{
Plots the locations of the origins of scalars from the postprocessor output
of bayestraits.
CURRENTLY WORKS ONLY FOR DELTAS.
}
